+++ personnal/src/main/java/cgy/controller/PositionController.java
@@ -5,6 +5,7 @@ import cgy.model.Page;
 import cgy.model.Position;
 import cgy.service.DepartmentService;
 import cgy.service.PositionService;
+import cgy.utils.CheckType;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
@@ -27,6 +28,8 @@ public class PositionController {
     //获得所有的职位信息
     @RequestMapping("getPositions")
     public String getPositions(Position position, HttpServletRequest request) {
+        boolean type = CheckType.getType(request);
+        if (type == false) return "jsp/login";
         String pageNoStr = request.getParameter("pageNo");
         int pageNo = 1;
         if (pageNoStr != null && !"".equals(pageNoStr)) {
@@ -65,6 +68,8 @@ public class PositionController {
     //添加一个职位
     @RequestMapping("addPosition")
     public String addPosition(Position position, HttpServletRequest request) {
+        boolean type = CheckType.getType(request);
+        if (type == false) return "jsp/login";
         System.out.println(position + "1");
         request.setAttribute("position", position);
         boolean insertPosition = positionService.insertPosition(position);
@@ -75,6 +80,8 @@ public class PositionController {
     //修改职位信息
     @RequestMapping("updatePosition")
     public String updatePosition(Position position, HttpServletRequest request) {
+        boolean type = CheckType.getType(request);
+        if (type == false) return "jsp/login";
         request.setAttribute("position", position);
         boolean updatePosition = positionService.updatePosition(position);
         request.setAttribute("updatePosition", updatePosition);
@@ -84,6 +91,8 @@ public class PositionController {
     //删除职位信息
     @RequestMapping("deleteposition")
     public String deleteposition(Position position, HttpServletRequest request) {
+        boolean type = CheckType.getType(request);
+        if (type == false) return "jsp/login";
         request.setAttribute("position", position);
         boolean deletePosition = positionService.deletePosition(position.getPos_id());
         request.setAttribute("deletePosition", deletePosition);
@@ -93,6 +102,8 @@ public class PositionController {
     //去给职位下的人分配培训
     @RequestMapping("toChoosePos")
     public String toChoosePos(int t_id, HttpServletRequest request) {
+        boolean type = CheckType.getType(request);
+        if (type == false) return "jsp/login";
         request.setAttribute("t_id", t_id);
         List<Department> departs = departmentService.getDeparts();
         request.setAttribute("departs", departs);
@@ -103,6 +114,8 @@ public class PositionController {
 
     @RequestMapping("toaddposition")
     public String toaddposition(HttpServletRequest request) {
+        boolean type = CheckType.getType(request);
+        if (type == false) return "jsp/login";
         List<Department> departments = departmentService.getDeparts();
         request.setAttribute("departments", departments);
         return "admin/positionadd";
