+++ personnal/src/main/java/cgy/controller/RecruitController.java
@@ -3,6 +3,7 @@ package cgy.controller;
 import cgy.model.*;
 import cgy.service.PositionService;
 import cgy.service.RecruitService;
+import cgy.utils.CheckType;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
 
@@ -40,6 +41,8 @@ public class RecruitController {
     //添加一条草稿
     @RequestMapping("addRecruit")
     public String addRecruit(Recruit recruit, HttpServletRequest request) {
+        boolean type = CheckType.getType(request);
+        if (type == false) return "jsp/login";
         Employee employee = (Employee) request.getSession().getAttribute("employeeNow");
         recruit.setRct_e_id(employee.getE_id());
         if ("null".equals(recruit.getRct_title())){
@@ -58,6 +61,8 @@ public class RecruitController {
     //删除一条草稿
     @RequestMapping("deleteRct")
     public String deleteRct(Recruit recruit, HttpServletRequest request) {
+        boolean type = CheckType.getType(request);
+        if (type == false) return "jsp/login";
         boolean deleteRct = recruitService.deleteRecruit(recruit.getRct_id());
         request.setAttribute("deleteRct",deleteRct);
         return "admin/recruit/recruitlist";
@@ -66,6 +71,8 @@ public class RecruitController {
     //更新一条草稿的状态
     @RequestMapping("updateRct")
     public String updateRct(Recruit recruit, HttpServletRequest request) {
+        boolean type = CheckType.getType(request);
+        if (type == false) return "jsp/login";
         recruit.setRct_publish_time(new Date());
         boolean updateRct = recruitService.updateRecruit(recruit);
         request.setAttribute("updateRct",updateRct);
@@ -82,6 +89,8 @@ public class RecruitController {
     //修改一条草稿
     @RequestMapping("updateRct2")
     public String updateRct2(Recruit recruit, HttpServletRequest request) {
+        boolean type = CheckType.getType(request);
+        if (type == false) return "jsp/login";
         boolean updateRct2 = recruitService.updateRecruit(recruit);
         request.setAttribute("updateRct",updateRct2);
         request.setAttribute("mark","");
